Move GetHoverAsync implementation to Workspaces
